+++ SpatCommon/Source/MatrixSlider.h
@@ -158,7 +158,7 @@ namespace Miam
                 // Displayable value
                 else
                 {
-                    float hue = (getValue()-GetMinVolume_dB()) / (GetMaxVolume_dB()-GetMinVolume_dB());
+                    double hue = (getValue()-GetMinVolume_dB()) / (GetMaxVolume_dB()-GetMinVolume_dB());
                     hue = (1.0-hue)*0.5 + 0.1;
                     uint8 alpha;
                     if (GetIsActive())
@@ -167,7 +167,7 @@ namespace Miam
                         alpha = 0x77;
                     
                     
-                    SetBackgroundColour(Colour(hue, 1.0, 0.7, alpha));
+                    SetBackgroundColour(Colour((float)hue, 1.0f, 0.7f, alpha));
                 }
                 
                 // In any case : possible value on next double click event
